pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-broker"
version = "0.1.0"
description = "OAuth token broker and reverse proxy for remote MCP server connections"
license = "Apache-2.0"
authors = [
    { name = "Firney", email = "dev@firney.com" },
]
keywords = ["mcp", "oauth", "broker", "proxy", "token", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Security",
    "Topic :: Internet :: Proxy Servers",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "httpx-sse",
    "itsdangerous",
    "pydantic",
    "pyyaml",
    "cryptography",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
    "pyright>=1.1.0",
    "pip-audit>=2.7.0",
]
twitter = ["xdk>=0.9.0,<1.0.0"]
all = ["mcp-broker[twitter]"]

[project.urls]
Homepage = "https://github.com/FirneyGroup/mcp-broker"
Repository = "https://github.com/FirneyGroup/mcp-broker"
Documentation = "https://github.com/FirneyGroup/mcp-broker#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "W", "F",
    "I",
    "B", "BLE",
    "C4", "C90", "SIM",
    "UP",
    "G",
    "S",
    "TID252",
    "PLR0913", "PLR0915", "PLR2004",
]
ignore = [
    "E501",
    "SIM108",
    "SIM117",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-args = 4
max-statements = 25

[tool.ruff.lint.isort]
known-first-party = ["broker", "connectors"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "PLR2004", "PLR0913", "B017"]
"**/test_*.py" = ["S101", "S105", "S106", "PLR2004", "PLR0913", "B017"]
"scripts/**/*.py" = ["S603", "S607"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["broker", "connectors"]
branch = true
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
fail_under = 65
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
